+++ vault/logical_system_test.go
@@ -12,12 +12,9 @@ import (
 
 func TestSystemBackend_RootPaths(t *testing.T) {
 	expected := []string{
-		"mounts/*",
 		"auth/*",
 		"remount",
 		"revoke-prefix/*",
-		"policy",
-		"policy/*",
 		"audit",
 		"audit/*",
 		"seal",
@@ -433,7 +430,7 @@ func TestSystemBackend_policyList(t *testing.T) {
 	}
 
 	exp := map[string]interface{}{
-		"keys": []string{"root"},
+		"keys": []string{"default", "root"},
 	}
 	if !reflect.DeepEqual(resp.Data, exp) {
 		t.Fatalf("got: %#v expect: %#v", resp.Data, exp)
@@ -485,7 +482,7 @@ func TestSystemBackend_policyCRUD(t *testing.T) {
 	}
 
 	exp = map[string]interface{}{
-		"keys": []string{"foo", "root"},
+		"keys": []string{"default", "foo", "root"},
 	}
 	if !reflect.DeepEqual(resp.Data, exp) {
 		t.Fatalf("got: %#v expect: %#v", resp.Data, exp)
@@ -519,7 +516,7 @@ func TestSystemBackend_policyCRUD(t *testing.T) {
 	}
 
 	exp = map[string]interface{}{
-		"keys": []string{"root"},
+		"keys": []string{"default", "root"},
 	}
 	if !reflect.DeepEqual(resp.Data, exp) {
 		t.Fatalf("got: %#v expect: %#v", resp.Data, exp)
@@ -682,7 +679,7 @@ func TestSystemBackend_rawWrite(t *testing.T) {
 	}
 
 	// Read the policy!
-	p, err := c.policy.GetPolicy("test")
+	p, err := c.policyStore.GetPolicy("test")
 	if err != nil {
 		t.Fatalf("err: %v", err)
 	}
@@ -709,7 +706,7 @@ func TestSystemBackend_rawDelete(t *testing.T) {
 
 	// set the policy!
 	p := &Policy{Name: "test"}
-	err := c.policy.SetPolicy(p)
+	err := c.policyStore.SetPolicy(p)
 	if err != nil {
 		t.Fatalf("err: %v", err)
 	}
@@ -725,8 +722,8 @@ func TestSystemBackend_rawDelete(t *testing.T) {
 	}
 
 	// Policy should be gone
-	c.policy.lru.Purge()
-	out, err := c.policy.GetPolicy("test")
+	c.policyStore.lru.Purge()
+	out, err := c.policyStore.GetPolicy("test")
 	if err != nil {
 		t.Fatalf("err: %v", err)
 	}