+++ vault/token_store.go
@@ -47,7 +47,7 @@ type TokenStore struct {
 
 	cubbyholeBackend *CubbyholeBackend
 
-	policyLookupFunc func() ([]string, error)
+	policyLookupFunc func(string) (*Policy, error)
 }
 
 // NewTokenStore is used to construct a token store that is
@@ -61,8 +61,8 @@ func NewTokenStore(c *Core, config *logical.BackendConfig) (*TokenStore, error) 
 		view: view,
 	}
 
-	if c.policy != nil {
-		t.policyLookupFunc = c.policy.ListPolicies
+	if c.policyStore != nil {
+		t.policyLookupFunc = c.policyStore.GetPolicy
 	}
 
 	// Setup the salt
@@ -89,6 +89,17 @@ func NewTokenStore(c *Core, config *logical.BackendConfig) (*TokenStore, error) 
 		},
 
 		Paths: []*framework.Path{
+			&framework.Path{
+				Pattern: "create-orphan$",
+
+				Callbacks: map[logical.Operation]framework.OperationFunc{
+					logical.WriteOperation: t.handleCreateOrphan,
+				},
+
+				HelpSynopsis:    strings.TrimSpace(tokenCreateOrphanHelp),
+				HelpDescription: strings.TrimSpace(tokenCreateOrphanHelp),
+			},
+
 			&framework.Path{
 				Pattern: "create$",
 
@@ -501,9 +512,23 @@ func (ts *TokenStore) revokeTreeSalted(saltedId string) error {
 	return nil
 }
 
-// handleCreate handles the auth/token/create path for creation of new tokens
+// handleCreate handles the auth/token/create path for creation of new orphan
+// tokens
+func (ts *TokenStore) handleCreateOrphan(
+	req *logical.Request, d *framework.FieldData) (*logical.Response, error) {
+	return ts.handleCreateCommon(req, d, true)
+}
+
+// handleCreate handles the auth/token/create path for creation of new non-orphan
+// tokens
 func (ts *TokenStore) handleCreate(
 	req *logical.Request, d *framework.FieldData) (*logical.Response, error) {
+	return ts.handleCreateCommon(req, d, false)
+}
+
+// handleCreateCommon handles the auth/token/create path for creation of new tokens
+func (ts *TokenStore) handleCreateCommon(
+	req *logical.Request, d *framework.FieldData, orphan bool) (*logical.Response, error) {
 	// Read the parent policy
 	parent, err := ts.Lookup(req.ClientToken)
 	if err != nil || parent == nil {
@@ -522,14 +547,15 @@ func (ts *TokenStore) handleCreate(
 
 	// Read and parse the fields
 	var data struct {
-		ID          string
-		Policies    []string
-		Metadata    map[string]string `mapstructure:"meta"`
-		NoParent    bool              `mapstructure:"no_parent"`
-		Lease       string
-		TTL         string
-		DisplayName string `mapstructure:"display_name"`
-		NumUses     int    `mapstructure:"num_uses"`
+		ID              string
+		Policies        []string
+		Metadata        map[string]string `mapstructure:"meta"`
+		NoParent        bool              `mapstructure:"no_parent"`
+		NoDefaultPolicy bool              `mapstructure:"no_default_policy"`
+		Lease           string
+		TTL             string
+		DisplayName     string `mapstructure:"display_name"`
+		NumUses         int    `mapstructure:"num_uses"`
 	}
 	if err := mapstructure.WeakDecode(req.Data, &data); err != nil {
 		return logical.ErrorResponse(fmt.Sprintf(
@@ -577,6 +603,9 @@ func (ts *TokenStore) handleCreate(
 		return logical.ErrorResponse("child policies must be subset of parent"), logical.ErrInvalidRequest
 	}
 	te.Policies = data.Policies
+	if !strListSubset(te.Policies, []string{"root"}) && !data.NoDefaultPolicy {
+		te.Policies = append(te.Policies, "default")
+	}
 
 	// Only allow an orphan token if the client has sudo policy
 	if data.NoParent {
@@ -586,6 +615,11 @@ func (ts *TokenStore) handleCreate(
 		}
 
 		te.Parent = ""
+	} else {
+		// This comes from create-orphan, which can be properly ACLd
+		if orphan {
+			te.Parent = ""
+		}
 	}
 
 	// Parse the TTL/lease if any
@@ -643,18 +677,13 @@ func (ts *TokenStore) handleCreate(
 	}
 
 	if ts.policyLookupFunc != nil {
-		availPolicies, err := ts.policyLookupFunc()
-		if err == nil {
-			policies := map[string]bool{}
-			if availPolicies != nil && len(availPolicies) > 0 {
-				for _, p := range availPolicies {
-					policies[p] = true
-				}
+		for _, p := range te.Policies {
+			policy, err := ts.policyLookupFunc(p)
+			if err != nil {
+				return logical.ErrorResponse(fmt.Sprintf("could not look up policy %s", p)), nil
 			}
-			for _, p := range te.Policies {
-				if !policies[p] {
-					resp.AddWarning(fmt.Sprintf("policy \"%s\" does not exist", p))
-				}
+			if policy == nil {
+				resp.AddWarning(fmt.Sprintf("policy \"%s\" does not exist", p))
 			}
 		}
 	}
@@ -775,10 +804,16 @@ func (ts *TokenStore) handleLookup(
 			"meta":          out.Meta,
 			"display_name":  out.DisplayName,
 			"num_uses":      out.NumUses,
+			"orphan":        false,
 			"creation_time": int(out.CreationTime),
 			"ttl":           int(out.TTL.Seconds()),
 		},
 	}
+
+	if out.Parent == "" {
+		resp.Data["orphan"] = true
+	}
+
 	return resp, nil
 }
 
@@ -839,6 +874,7 @@ Client tokens are used to identify a client and to allow Vault to associate poli
 which are enforced on every request. This backend also allows for generating sub-tokens as well
 as revocation of tokens. The tokens are renewable if associated with a lease.`
 	tokenCreateHelp       = `The token create path is used to create new tokens.`
+	tokenCreateOrphanHelp = `The token create path is used to create new orphan tokens.`
 	tokenLookupHelp       = `This endpoint will lookup a token and its properties.`
 	tokenRevokeHelp       = `This endpoint will delete the given token and all of its child tokens.`
 	tokenRevokeSelfHelp   = `This endpoint will delete the token used to call it and all of its child tokens.`